# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LinkClub API
  summary: Contrato objetivo de la API de LinkClub (diseño API-first, previo a la implementación completa).
  description: >
    Fuente única de verdad del contrato entre el backend de LinkClub y sus
    consumidores (app Flutter). Cubre las tres personas descritas en el C4 de
    contexto: Estudiante interesado, Miembro de club y Administrador de club.
    Este archivo describe el contrato TAL COMO DEBE SER, no necesariamente lo
    que ya está implementado; el estado de implementación por endpoint está en
    docs/api/contrato_api.md (sección 5).
  version: 1.1.0
  contact:
    name: Equipo LinkClub
    url: https://github.com/ISCOUTB/AS_202620_Clubs_UTB
servers:
  - url: http://localhost:8000
    description: Entorno local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /clubes/{club_id}/publicaciones:
    get:
      operationId: listarPublicaciones
      tags: [publicaciones]
      summary: Lista los avisos, encuestas y noticias de un club.
      security: []
      parameters:
        - $ref: "#/components/parameters/ClubId"
      responses:
        "200":
          description: Publicaciones del club.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Publicacion" }
    post:
      operationId: crearPublicacion
      tags: [publicaciones]
      summary: Crea un aviso, una encuesta o una noticia para el club.
      description: Requiere rol "miembro" o "administrador" en ese club.
      parameters:
        - $ref: "#/components/parameters/ClubId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CrearPublicacionRequest" }
      responses:
        "201":
          description: Publicación creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publicacion" }
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/SinPermiso"
        "422":
          $ref: "#/components/responses/DatosInvalidos"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión emitido por Supabase Auth (restricción T4).
  parameters:
    ClubId:
      name: club_id
      in: path
      required: true
      description: Identificador del club.
      schema: { type: string }
...
  responses:
    NoAutenticado:
      description: Falta un token válido de Supabase Auth.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    SinPermiso:
      description: El usuario autenticado no tiene el rol requerido para esta acción.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    DatosInvalidos:
      description: El cuerpo de la petición no cumple las reglas del recurso.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
      additionalProperties: false
...
    TipoPublicacion:
      type: string
      description: >
        Enum extensible: agregar un nuevo tipo no debe requerir modificar más
        de 3 módulos (escenario de cambio C1).
      enum: [aviso, encuesta, noticia]
    CrearPublicacionRequest:
      type: object
      required: [titulo, contenido]
      properties:
        titulo: { type: string, minLength: 1 }
        contenido: { type: string, minLength: 1 }
        tipo:
          allOf: [{ $ref: "#/components/schemas/TipoPublicacion" }]
          default: aviso
        opciones:
          type: [array, "null"]
          items: { type: string }
          description: Obligatorio y no vacío cuando tipo = "encuesta".
      additionalProperties: false
      if:
        properties: { tipo: { const: encuesta } }
        required: [tipo]
      then:
        required: [opciones]
        properties:
          opciones: { type: array, minItems: 1, items: { type: string } }
    Publicacion:
      type: object
      required: [id, club_id, titulo, contenido, tipo, creado_en]
      properties:
        id: { type: string, format: uuid }
        club_id: { type: string }
        autor_id: { type: string }
        titulo: { type: string }
        contenido: { type: string }
        tipo: { $ref: "#/components/schemas/TipoPublicacion" }
        opciones:
          type: [array, "null"]
          items: { type: string }
        creado_en: { type: string, format: date-time }
      additionalProperties: false